Improve picture_snapper helper node
Added a launch file to start up only the picture_snapper node; requires
some camera to be running. A small guide was added to the top of the
folder that describes this process. This node is also much more
interactive, allowing you to specify a save location using the
image_save_dir param. The node logs more detailed information including
an absolute path to the saved picture.