Improve static_transform_pub helper node
Added a service to the node that allows the user to save the static
transforms at any time, rather than only at node shutdown. Added
more params to the static_transform_pub launch file that allow for
more control over where the file is saved. The node provides a better
way to specify the file name and path, creating directories in the path
if none exist. Added more logging features to the node including the
absolute path to the transforms file when saved and loaded.