[perception_toolboxes] Added landmark module
Landmarks will be used to keep track of locations of interest within a
map. The landmark module captures landmark poses and adds them to
the static tf tree. This commit also adds some tools to ensure
observation validity through the new geometry module. For now, this
module only contains a method to check if quaternions are valid (since
I was not able to find that method anywhere else).