[perception_toolboxes] Improved helper nodes
Added the ability to toggle saving on the static_trans_pub node. This
was done to give the user the option to not save any static tf sent to
this node.
Refactored apriltag helper node. Added a helper method to trigger the
apriltag single image service rather than have it contained within a
more specific method, exposing it to anything that wants to use the
service. Added a method to set the list of valid tags (the list of tags
that will be seen by the node) to ids other than the defaults. Added a
method that detects any number of tags and returns a list of poses and
the corresponding list of ids.