[landmark_modules] Landmark Manager Improvements
Removed references to landmark manager GUI. Added a feature where
landmarks that already exist will be highlighted in the CLI. Added a
parameter that tracks the standalone tags and sets it as the list of
valid tags. Added a launch file to make it easier to specify params.
Minor rewording of prompts. Fixed a bug where new landmarks would not
be added to the manager collection properly. Fixed a bug where the
params were not loaded properly.