[xs_modules] Add timeout and message to services
Adds a try/except block to the InterbotixRobotXSCore class when waiting
for the services, as well as timeouts to the wait_for_service functions.
If the functions timeout, an exception is raised and caught. The error
message is printed out and a line stating that the robot can't be found
and asks the user to check if they entered the right robot_model param.

This was done to solve an issue where users couldn't figure out why
their Python API scripts were hanging.

[skip ci]