[xs_rviz] Fix build errors
Catkin built packages in an order that would prevent service types from
being found before the xs_rviz package is built, leading to build
errors. This commit should fix this by changing the location of the
catkin_build() directive in CMakeLists and by changing the service type
to the service call rather than the service response.