[xs_modules] Add timeout to MATLAB core
Adds a try/except block to the InterbotixRobotXSCore class when waiting
for the services, as well as timeouts to the wait_for_service functions.
If the functions timeout, an exception is raised and caught. The error
message is printed out and a line stating that the robot can't be found
and asks the user to check if they entered the right robot_name param.

This was done to match the changes made in 638d37f.